Stake Chart with endpoint
