+++ src/pages/bond/overview/index.tsx
@@ -7,15 +7,38 @@ import Header from 'common/components/Bonds/BondsSummaryHeader/Header'
 // import { BondEvents } from 'modules/BondEvents/BondEvents.container'
 import { selectLocationProps } from 'modules/Router/router.selector'
 import { getTransactionsByBondDID } from 'modules/BondModules/bond/bond.actions'
+import { RootState } from 'common/redux/types'
+import { getAccount, getTransactionsByAsset } from 'modules/Account/Account.actions'
 
 export const Overview: FunctionComponent<any> = ({ match }) => {
   const dispatch = useDispatch()
   const [selectedHeader, setSelectedHeader] = useState('price')
   const location: any = useSelector(selectLocationProps)
+  const {
+    address: accountAddress,
+    balances,
+  } = useSelector((state: RootState) => state.account)
 
   useEffect(() => {
     dispatch(getTransactionsByBondDID())
   }, [dispatch])
+  
+  useEffect(() => {
+    accountAddress && dispatch(getAccount(accountAddress))
+    // eslint-disable-next-line
+  }, [accountAddress])
+
+  useEffect(() => {
+    if (balances.length > 0) {
+      dispatch(
+        getTransactionsByAsset(
+          accountAddress,
+          balances.map((balance) => balance.denom),
+        ),
+      )
+    }
+    // eslint-disable-next-line
+  }, [balances])
 
   const projectPublic =
     location.state && location.state.projectPublic
